feat: refactor store structure; move store configuration to a new file and update imports in login components
