+++ frontend/src/features/store.ts
@@ -1,6 +1,6 @@
 // src/store.ts
 import { configureStore } from '@reduxjs/toolkit';
-import authReducer from '../auth/authSlice';
+import authReducer from './auth/authSlice';
 
 export const store = configureStore({
     reducer: {
