+++ frontend/src/main.tsx
@@ -2,7 +2,7 @@ import { createRoot } from 'react-dom/client'
 import { Provider } from "react-redux";
 import './index.css'
 import App from './App.tsx'
-import { store } from './features/app/store.ts';
+import { store } from './features/store.ts';
 import { ThemeProvider } from "./components/ui/theme-provider.tsx"
 
 createRoot(document.getElementById('root')!).render(
